feat(budget): create hard coded budget card #5
